requirements.txt
altair
pydeck
streamlit
streamlit
pandas
numpy
openai==1.6.1
openpyxl==3.1.2
farm-haystack[inference]
torch==1.10.0
transformers==4.35.2
sentence-transformers